+++ internal/resource/common/attributesources/attribute_sources_state.go
@@ -12,18 +12,22 @@ import (
 	"github.com/pingidentity/terraform-provider-pingfederate/internal/resource/common/resourcelink"
 )
 
-func CommonAttributeSourceAttrType() map[string]attr.Type {
+func CommonAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation bool) map[string]attr.Type {
 	commonAttrSourceAttrType := map[string]attr.Type{}
 	commonAttrSourceAttrType["type"] = types.StringType
-	commonAttrSourceAttrType["data_store_ref"] = types.ObjectType{AttrTypes: resourcelink.AttrType()}
+	if includeDataStoreRefResourceLinkLocation {
+		commonAttrSourceAttrType["data_store_ref"] = types.ObjectType{AttrTypes: resourcelink.AttrType()}
+	} else {
+		commonAttrSourceAttrType["data_store_ref"] = types.ObjectType{AttrTypes: resourcelink.AttrTypeNoLocation()}
+	}
 	commonAttrSourceAttrType["id"] = types.StringType
 	commonAttrSourceAttrType["description"] = types.StringType
 	commonAttrSourceAttrType["attribute_contract_fulfillment"] = attributecontractfulfillment.MapType()
 	return commonAttrSourceAttrType
 }
 
-func CustomAttributeSourceAttrType() map[string]attr.Type {
-	customAttrSourceAttrType := CommonAttributeSourceAttrType()
+func CustomAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation bool) map[string]attr.Type {
+	customAttrSourceAttrType := CommonAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation)
 	customAttrSourceAttrType["filter_fields"] = types.ListType{ElemType: types.ObjectType{
 		AttrTypes: map[string]attr.Type{
 			"value": types.StringType,
@@ -33,8 +37,8 @@ func CustomAttributeSourceAttrType() map[string]attr.Type {
 	return customAttrSourceAttrType
 }
 
-func JdbcAttributeSourceAttrType() map[string]attr.Type {
-	jdbcAttributeSourceAttrType := CommonAttributeSourceAttrType()
+func JdbcAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation bool) map[string]attr.Type {
+	jdbcAttributeSourceAttrType := CommonAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation)
 	jdbcAttributeSourceAttrType["schema"] = types.StringType
 	jdbcAttributeSourceAttrType["table"] = types.StringType
 	jdbcAttributeSourceAttrType["column_names"] = types.ListType{ElemType: types.StringType}
@@ -42,8 +46,8 @@ func JdbcAttributeSourceAttrType() map[string]attr.Type {
 	return jdbcAttributeSourceAttrType
 }
 
-func LdapAttributeSourceAttrType() map[string]attr.Type {
-	ldapAttrSourceAttrType := CommonAttributeSourceAttrType()
+func LdapAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation bool) map[string]attr.Type {
+	ldapAttrSourceAttrType := CommonAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation)
 	ldapAttrSourceAttrType["base_dn"] = types.StringType
 	ldapAttrSourceAttrType["search_scope"] = types.StringType
 	ldapAttrSourceAttrType["search_filter"] = types.StringType
@@ -59,35 +63,35 @@ func LdapAttributeSourceAttrType() map[string]attr.Type {
 	return ldapAttrSourceAttrType
 }
 
-func AttrType() map[string]attr.Type {
+func AttrType(includeDataStoreRefResourceLinkLocation bool) map[string]attr.Type {
 	return map[string]attr.Type{
 		"attribute_sources": types.ListType{
 			ElemType: types.ObjectType{
-				AttrTypes: ElemAttrType(),
+				AttrTypes: ElemAttrType(includeDataStoreRefResourceLinkLocation),
 			},
 		},
 	}
 }
 
-func ElemAttrType() map[string]attr.Type {
+func ElemAttrType(includeDataStoreRefResourceLinkLocation bool) map[string]attr.Type {
 	return map[string]attr.Type{
 		"custom_attribute_source": types.ObjectType{
-			AttrTypes: CustomAttributeSourceAttrType(),
+			AttrTypes: CustomAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation),
 		},
 		"jdbc_attribute_source": types.ObjectType{
-			AttrTypes: JdbcAttributeSourceAttrType(),
+			AttrTypes: JdbcAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation),
 		},
 		"ldap_attribute_source": types.ObjectType{
-			AttrTypes: LdapAttributeSourceAttrType(),
+			AttrTypes: LdapAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation),
 		},
 	}
 }
 
-func ToState(con context.Context, attributeSourcesFromClient []client.AttributeSourceAggregation) (basetypes.ListValue, diag.Diagnostics) {
+func ToState(con context.Context, attributeSourcesFromClient []client.AttributeSourceAggregation, includeDataStoreRefResourceLinkLocation bool) (basetypes.ListValue, diag.Diagnostics) {
 	var diags diag.Diagnostics
-	var customAttrSourceAttrTypes = CustomAttributeSourceAttrType()
-	var jdbcAttrSourceAttrTypes = JdbcAttributeSourceAttrType()
-	var ldapAttrSourceAttrTypes = LdapAttributeSourceAttrType()
+	var customAttrSourceAttrTypes = CustomAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation)
+	var jdbcAttrSourceAttrTypes = JdbcAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation)
+	var ldapAttrSourceAttrTypes = LdapAttributeSourceAttrType(includeDataStoreRefResourceLinkLocation)
 	var valueFromDiags diag.Diagnostics
 
 	// Build attribute_sources value
@@ -101,7 +105,14 @@ func ToState(con context.Context, attributeSourcesFromClient []client.AttributeS
 			diags.Append(valueFromDiags...)
 
 			customAttrSourceValues["type"] = types.StringValue("CUSTOM")
-			customAttrSourceValues["data_store_ref"], valueFromDiags = types.ObjectValueFrom(con, resourcelink.AttrType(), attrSource.CustomAttributeSource.DataStoreRef)
+			if includeDataStoreRefResourceLinkLocation {
+				customAttrSourceValues["data_store_ref"], valueFromDiags = types.ObjectValueFrom(con, resourcelink.AttrType(), attrSource.CustomAttributeSource.DataStoreRef)
+			} else {
+				customDataStoreRefAttrValue := map[string]attr.Value{
+					"id": types.StringValue(attrSource.CustomAttributeSource.DataStoreRef.Id),
+				}
+				customAttrSourceValues["data_store_ref"], valueFromDiags = types.ObjectValue(resourcelink.AttrTypeNoLocation(), customDataStoreRefAttrValue)
+			}
 			diags.Append(valueFromDiags...)
 			customAttrSourceValues["id"] = types.StringPointerValue(attrSource.CustomAttributeSource.Id)
 			customAttrSourceValues["description"] = types.StringPointerValue(attrSource.CustomAttributeSource.Description)
@@ -120,7 +131,14 @@ func ToState(con context.Context, attributeSourcesFromClient []client.AttributeS
 			diags.Append(valueFromDiags...)
 			jdbcAttrSourceValues["filter"] = types.StringValue(attrSource.JdbcAttributeSource.Filter)
 			jdbcAttrSourceValues["type"] = types.StringValue("JDBC")
-			jdbcAttrSourceValues["data_store_ref"], valueFromDiags = types.ObjectValueFrom(con, resourcelink.AttrType(), attrSource.JdbcAttributeSource.DataStoreRef)
+			if includeDataStoreRefResourceLinkLocation {
+				jdbcAttrSourceValues["data_store_ref"], valueFromDiags = types.ObjectValueFrom(con, resourcelink.AttrType(), attrSource.JdbcAttributeSource.DataStoreRef)
+			} else {
+				jdbcDataStoreRefAttrValue := map[string]attr.Value{
+					"id": types.StringValue(attrSource.JdbcAttributeSource.DataStoreRef.Id),
+				}
+				jdbcAttrSourceValues["data_store_ref"], valueFromDiags = types.ObjectValue(resourcelink.AttrTypeNoLocation(), jdbcDataStoreRefAttrValue)
+			}
 			diags.Append(valueFromDiags...)
 			jdbcAttrSourceValues["id"] = types.StringPointerValue(attrSource.JdbcAttributeSource.Id)
 			jdbcAttrSourceValues["description"] = types.StringPointerValue(attrSource.JdbcAttributeSource.Description)
@@ -146,7 +164,14 @@ func ToState(con context.Context, attributeSourcesFromClient []client.AttributeS
 			}
 			ldapAttrSourceValues["member_of_nested_group"] = types.BoolPointerValue(attrSource.LdapAttributeSource.MemberOfNestedGroup)
 			ldapAttrSourceValues["type"] = types.StringValue(attrSource.LdapAttributeSource.Type)
-			ldapAttrSourceValues["data_store_ref"], valueFromDiags = types.ObjectValueFrom(con, resourcelink.AttrType(), attrSource.LdapAttributeSource.DataStoreRef)
+			if includeDataStoreRefResourceLinkLocation {
+				ldapAttrSourceValues["data_store_ref"], valueFromDiags = types.ObjectValueFrom(con, resourcelink.AttrType(), attrSource.LdapAttributeSource.DataStoreRef)
+			} else {
+				ldapDataStoreRefAttrValue := map[string]attr.Value{
+					"id": types.StringValue(attrSource.LdapAttributeSource.DataStoreRef.Id),
+				}
+				ldapAttrSourceValues["data_store_ref"], valueFromDiags = types.ObjectValue(resourcelink.AttrTypeNoLocation(), ldapDataStoreRefAttrValue)
+			}
 			diags.Append(valueFromDiags...)
 			ldapAttrSourceValues["id"] = types.StringPointerValue(attrSource.LdapAttributeSource.Id)
 			ldapAttrSourceValues["description"] = types.StringPointerValue(attrSource.LdapAttributeSource.Description)
@@ -157,11 +182,11 @@ func ToState(con context.Context, attributeSourcesFromClient []client.AttributeS
 		} else {
 			attrSourceValues["ldap_attribute_source"] = types.ObjectNull(ldapAttrSourceAttrTypes)
 		}
-		attrSourceElement, valueFromDiags := types.ObjectValue(ElemAttrType(), attrSourceValues)
+		attrSourceElement, valueFromDiags := types.ObjectValue(ElemAttrType(includeDataStoreRefResourceLinkLocation), attrSourceValues)
 		diags.Append(valueFromDiags...)
 		attrSourceElements = append(attrSourceElements, attrSourceElement)
 	}
-	attrToState, valueFromDiags := types.ListValue(types.ObjectType{AttrTypes: ElemAttrType()}, attrSourceElements)
+	attrToState, valueFromDiags := types.ListValue(types.ObjectType{AttrTypes: ElemAttrType(includeDataStoreRefResourceLinkLocation)}, attrSourceElements)
 	diags.Append(valueFromDiags...)
 	return attrToState, diags
 }