+++ internal/resource/config/passwordcredentialvalidator/password_credential_validator_resource.go
@@ -106,7 +106,7 @@ func (r *passwordCredentialValidatorResource) Schema(ctx context.Context, req re
 						},
 					},
 					"inherited": schema.BoolAttribute{
-						DeprecationMessage: "This field is now deprecated and will be removed in the next release.",
+						DeprecationMessage: "This field is now deprecated and will be removed in a future release.",
 						Description:        "Whether this attribute contract is inherited from its parent instance. If true, the rest of the properties in this model become read-only. The default value is false.",
 						Optional:           true,
 						Computed:           true,